[AIRFLOW-1005][AIRFLOW-940][AIRFLOW-1516] Error handling var decrypt to prevent braking VariableView (#98)
* [AIRFLOW-1005] Improve Airflow startup time

Airflow’s startup time can be reduced by 50% by
deferring imports of Cryptography (and relatedly,
not generating Fernet keys unless we have to) and
Alembic.

* [AIRFLOW-940] Handle error on variable decrypt

Invalid variables could break the variable view by
unhandled
InvalidToken exception from Fernet.

This commit converts the fernet error into an
AirflowException, given
that fernet is loaded dynamically. Also, the
exception is handled in
the VariableView by showing the token "INVALID" in
the UI render.

Closes #2510 from edgarRd/erod-error-handling-var-
decrypt

* [AIRFLOW-1516] Fix error handling getting fernet

There were unhandled cases for exceptions when
importing fernet in
models.py. This seems to be a remanent of a
previous refactor,
replacing logic that would depend on the
definition of a global variable
for Fernet if it was imported correctly.

Generally catching all exceptions from get_fernet
function, given that
other functions are already handling it that way
and the only
error handling case here is to not use encryption.

Closes #2527 from edgarRd/erod-fernet-error-
handling